+++ eQuiz/modules/eQuiz.Web/Areas/Admin/Models/SelectQuestion.cs
@@ -9,7 +9,7 @@ namespace eQuiz.Web.Areas.Admin.Models
     {
         public IEnumerable<TestAnswer> QuestionVariants;
 
-        public SelectQuestion(int id, byte maxScore, int? userScore, string questionText, IEnumerable<TestAnswer> questionVariants, short? order, bool wasChanged, bool wasNull)
+        public SelectQuestion(int id, byte maxScore, double? userScore, string questionText, IEnumerable<TestAnswer> questionVariants, short? order, bool wasChanged, bool wasNull)
             : base(id, maxScore, userScore, questionText, order, wasChanged, wasNull)
         {
             base.Type = "Select";
